[NaCl SDK] Enable linux host build for nacl_io and nacl_io_tests.
The process of doing this already found one serious bug
in lseek and will enable us to find many more via asan and
other sanitisers which run on the linux host envionment.

R=binji@chromium.org, binji

Review URL: https://codereview.chromium.org/73083005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@235753 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 9f4a33d48f6916c7190b1de572a3fc857f684394

+++ native_client_sdk/src/libraries/nacl_io/fifo_interface.h
@@ -23,8 +23,8 @@ class FIFOInterface {
   virtual bool IsFull() = 0;
   virtual bool Resize(size_t len) = 0;
 
-  virtual uint32_t ReadAvailable() = 0;
-  virtual uint32_t WriteAvailable() = 0;
+  virtual size_t ReadAvailable() = 0;
+  virtual size_t WriteAvailable() = 0;
 };
 
 }  // namespace nacl_io
